refactor: DataFrameProvider now placed in one module
Fixes: #33
